refactor: renamed SearchForm component and simplified Card component markup
